Move and rename ls_controller
