+++ 08.ls_object/bin/ls
@@ -1,6 +1,6 @@
 #!/usr/bin/env ruby
 # frozen_string_literal: true
 
-require_relative '../lib/controllers/ls_controller'
+require_relative '../lib/ls_app'
 
-LsController.new.main
+LsApp.new.main
