+++ 08.ls_object/lib/ls_app.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
 require_relative 'params'
-require_relative '../models/ls_file'
-require_relative '../views/short_format'
-require_relative '../views/long_format'
+require_relative 'models/ls_file'
+require_relative 'views/short_format'
+require_relative 'views/long_format'
 
-class LsController
+class LsApp
   def main(params = Params.new(ARGV))
     ls_files = LsFile.all(params)
     if params.long_format?
