+++ 08.ls_object/lib/params.rb
